Create README.md
Mainly three algorithms are used in system they are as follows:
•	Natural Language  Processing(NLP) algorithm
Natural language processing (NLP) is a field of artificial intelligence in which computers analyze, understand, and derive meaning from human language in a smart and useful way. By utilizing NLP, developers can organize and structure knowledge to perform tasks such as automatic summarization, translation, named entity recognition, relationship extraction, sentiment analysis, speech recognition, and topic segmentation. We use this algorithm for cleaning the data. It tokenize the sentence to word. It then removes stop words and special characters from sentence so that processing could be easy.  

•	Long short-term memory(LSTM)
It is an artificial recurrent neural network (RNN) architecture[1] used in the field of deep learning. Unlike standard feedforward neural networks, LSTM has feedback connections. It can not only process single data points (such as images), but also entire sequences of data (such as speech or video).After processing using NLP we load a pre-trained model using word2vec .Then we convert conversation sentences to x_vec and y_vec vectors they are ten coverted to numeric format using KeyedVector and dumped in pickle file it completes the preprocessing. LSTM is used for training our model first we extract the pickle file and give it to LSTM neural network where note are trained it can remember information for longer time. The  activation function used for LSTM is sigmoid and 500 epoches.   

•	Apriori algorithm 
Apriori is an algorithm for frequent item set mining and association rule learning over relational databases. It proceeds by identifying the frequent individual items in the database and extending them to larger and larger item sets as long as those item sets appear sufficiently often in the database. The frequent item sets determined by Apriori can be used to determine association rules which highlight general trends in the database this has applications in domains such as market basket analysis. We use this algorithm for recommending food to user it returns top 10 items that were purchase the most by customer from which bot recommends one of them in random way.